% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCM_download_data.R
\name{gcm_download_data}
\alias{gcm_download_data}
\title{Download daily GCM data from NCCS THREDDS NEX-GDDP-CMIP6.}
\usage{
gcm_download_data(
  location,
  model = NULL,
  scenario,
  variable,
  years,
  roi,
  version = NULL,
  method = "curl"
)
}
\arguments{
\item{location}{Select the location for data downloading.}

\item{model}{Model name. If NULL, all available models will be selected.}

\item{scenario}{CMIP6 scenario ('historical','ssp126','ssp245', 'ssp370', or 'ssp585).}

\item{variable}{Variable ('hurs','huss','pr','rlds','rsds','sfcWind','tas','tasmax', or 'tasmin').}

\item{years}{Data years (1950 - 2014 for 'historical' and 2015 - 2100 for 'ssp126', 'ssp245', 'ssp370' and 'ssp585').}

\item{roi}{Extension for sub-setting gridded data (west, east, south, north). If NULL, original extension is considered.}

\item{version}{Select file version. Use NULL for first version, or 'v1.1' and 'v1.2'. Please read technical notes in NEX website.}

\item{method}{Method to be used for downloading files. Current download methods are 'internal', 'wininet' (Windows only), 'libcurl', 'wget' and 'curl'. The 'curl' method is recommended for Windows users.}
}
\value{
CMIP6 daily data in netCDF format.
}
\description{
Download daily GCM data from NCCS THREDDS NEX-GDDP-CMIP6.
}
\examples{
# Load package
require(RClimChange)

# Download daily precipitation from the BCC-CSM2-MR model
# from historical period
gcm_download_data(location=getwd(),
                  model='BCC-CSM2-MR',
                  scenario='historical',
                  variable='pr',
                  years=1990,
                  version='v1.1',
                  roi=c(-86,-66,-20,2),
                  method='curl')

# from ssp585 scenario
gcm_download_data(location=getwd(),
                  model='BCC-CSM2-MR',
                  scenario='historical',
                  variable='pr',
                  years=1991:1992,
                  version='v1.1',
                  roi=c(-86,-66,-20,2),
                  method='curl')

}
