% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCM_extract_region.R
\name{GCM_extract_region}
\alias{GCM_extract_region}
\title{Read subset and built a 5D array with netCDF GCM's files}
\usage{
GCM_extract_region(Path.GCM, Region, Var)
}
\arguments{
\item{Path.GCM}{Directory where GCM netCDF files are located.}

\item{Region}{Lat/Lon ranges for subsetting data (MinLon, MaxLon, MinLat, MaxLat).}

\item{Var}{Variable of intereset to be extracted (Pp, Tmax or Tmin).}
}
\value{
5D array [lon,lat,day,year,model] with GCM subset data.
}
\description{
Read subset and built a 5D array with netCDF GCM's files
}
