% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCM_spatial_down.R
\name{GCM_spatial_down}
\alias{GCM_spatial_down}
\title{Extract a downscaled GCM data for a point of interest}
\usage{
GCM_spatial_down(Input, RegLon, RegLat, XYcoords, YearIni, b = 2)
}
\arguments{
\item{Input}{5D array [lon,lat,day,year,model] with GCM subset data.}

\item{RegLon}{Vector of longitudes for the study area.}

\item{RegLat}{Vector of latitudes for the study area.}

\item{XYcoords}{Station coordinates (lon, lat) for downscaling data.}

\item{YearIni}{Initial year of data (to compute leap years).}

\item{b}{IDW exponent (2 as default).}
}
\value{
GCM time series for a point of interest.
}
\description{
Extract a downscaled GCM data for a point of interest
}
