% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCM_download_data.R
\name{GCM_download_data}
\alias{GCM_download_data}
\title{Download and reshape netCDF files from NASA's NEX-GDDP dataset.}
\usage{
GCM_download_data(FileURLs, Variable = "pr", RangeLat, RangeLon)
}
\arguments{
\item{FileURLs}{Text file with NEX-GDDP URLs to be downloaded.}

\item{Variable}{Variable of interest to be downloaded (pr, tasmax or tasmin). 'pr' as default.}

\item{RangeLat}{Range (min,max) of latitudes to subset data.}

\item{RangeLon}{Range (min,max) of longitudes to subset data.}
}
\value{
NetCDF files extracted for an study region and variable of interest.
}
\description{
Download and reshape netCDF files from NASA's NEX-GDDP dataset.
}
\examples{
require(RClimChange)

GCM_download_data(FileURLs='pr_rcp4.5.txt', #Download precipitation for RCP 4.5
                  Variable='pr',
                  RangeLat=c(-20.5,3.2),
                  RangeLon=c(-82.5,-64.4)) #Extent for Peru

}
