% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCM_download_data.R
\name{GCM_download_data}
\alias{GCM_download_data}
\title{Download and cut netCDF files from NASA's NEX-GDDP GCM files}
\usage{
GCM_download_data(Path, FileURLs, Var, RangeLat, RangeLon)
}
\arguments{
\item{Path}{Directory where FileURLs is located}

\item{FileURLs}{CSV file with urls to download data}

\item{Var}{Variable to download data ('Pp', 'Tmax or 'Tmin')}

\item{RangeLat}{Limits of latitudes to subset data}

\item{RangeLon}{Limits of longitudes to subset data}
}
\value{
Downloaded netCDF files for a region and variable of interes
}
\description{
Download and cut netCDF files from NASA's NEX-GDDP GCM files
}
