% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcm_download_data.R
\name{gcm_download_data}
\alias{gcm_download_data}
\title{Download CMIP6 daily data from NCCS THREDDS NEX-GDDP-CMIP6 (https://www.nccs.nasa.gov/sites/default/files/NEX-GDDP-CMIP6-Tech_Note.pdf)}
\usage{
gcm_download_data(
  location,
  model = NULL,
  scenario,
  variable,
  years,
  roi,
  method = "curl"
)
}
\arguments{
\item{location}{Work directory to store downloaded data.}

\item{model}{Model names to download. If NULL, all available models will be selected.}

\item{scenario}{Choose the scenario to be downloaded ('historical','ssp245', or 'ssp585).}

\item{variable}{Choose the variable to be downloaded ('hurs','huss','pr','rlds','rsds','sfcWind','tas','tasmax', or 'tasmin').}

\item{years}{Choose data years to be downloaded  (1950-2014 for 'historical' and 2015-2100 for 'ssp245' and 'ssp585').}

\item{roi}{Vector of coordinates for subsetting data (xmin, xmax, ymin, ymax). If NULL, original extension data will be downloaded.}

\item{method}{Method to be used for downloading files. Current download methods are 'internal', 'wininet' (Windows only), 'libcurl', 'wget' and 'curl'. The 'curl' method is recommended for Windows users.}
}
\value{
CMIP6 daily data (in netCDF format).
}
\description{
Download CMIP6 daily data from NCCS THREDDS NEX-GDDP-CMIP6 (https://www.nccs.nasa.gov/sites/default/files/NEX-GDDP-CMIP6-Tech_Note.pdf)
}
\examples{
# Load package
require(RClimChange)

# Download daily precipitation (in mm/d) from the BCC-CSM2-MR model; for 1990 and the Peruvian domain
gcm_download_data(location=getwd(),
                  model='BCC-CSM2-MR',
                  scenario='historical',
                  variable='pr',
                  years=1990,
                  roi=c(-86,-66,-20,2),
                  method='curl')

}
