% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcm_download_data.R
\name{gcm_download_data}
\alias{gcm_download_data}
\title{Download CMIP6 daily data from NCCS THREDDS NEX-GDDP-CMIP6}
\usage{
gcm_download_data(
  location,
  model = NULL,
  scenario,
  variable,
  years,
  roi,
  method = "libcurl"
)
}
\arguments{
\item{location}{Work directory to store downloaded data.}

\item{model}{Model names from "https://ds.nccs.nasa.gov/thredds/catalog/AMES/NEX/GDDP-CMIP6/catalog.html". If NULL, all 35 models will be downloaded.}

\item{scenario}{Choose the scenario to be downloaded ('historical','ssp245','ssp585). More information in "https://www.nccs.nasa.gov/sites/default/files/NEX-GDDP-CMIP6-Tech_Note.pdf".}

\item{variable}{Choose the variable to be downloaded ('pr':precipitation,'tas': mean air temperature ,'tasmin': minimum air temperature,'tasmax': maximum air temperature). 'pr' as default.}

\item{years}{Choose data years to be downloaded  (from 1950 to2014 for the 'historical' scenario, and from 2015 to 2100 for 'ssp245' and 'ssp585' scenarios).}

\item{roi}{Region of interest coordinates for subsetting data. Please insert c(xmin, xmax, ymin, ymax). If NULL, data with the original extension will be downloaded.}

\item{method}{Method to be used for downloading files. Current download methods are 'internal', 'wininet' (Windows only), 'libcurl', 'wget' and 'curl'.'libcurl' as default.}
}
\value{
CMIP6 daily data (in netCDF format) for the region of interest.
}
\description{
Download CMIP6 daily data from NCCS THREDDS NEX-GDDP-CMIP6
}
\examples{
require(RClimChange)

# Download daily precipitation (in mm/d) from the BCC-CSM2-MR model for 1990 and the Peruvian domain
gcm_download_data <- function(location=getwd(),
                              model='BCC-CSM2-MR',
                              scenario='historical',
                              variable='pr',
                              years=1990,
                              roi=c(-86,-66,-20,2), # extent for the Peruvian domain
                              method='libcurl')

}
