% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCM_download_data.R
\name{GCM_download_data}
\alias{GCM_download_data}
\title{Download and reshape netCDF files from NASA's NEX-GDDP dataset.}
\usage{
GCM_download_data(FileURLs, Var, RangeLat, RangeLon)
}
\arguments{
\item{FileURLs}{Text file with NEX-GDDP URLs to be downloaded.}

\item{Var}{Variable of interest to be downloaded (Pp, Tmax or Tmin).}

\item{RangeLat}{Range (min,max) of latitudes to subset data.}

\item{RangeLon}{Range (min,max) of longitudes to subset data.}
}
\value{
NetCDF files extracted for an study region and variable of interest.
}
\description{
Download and reshape netCDF files from NASA's NEX-GDDP dataset.
}
