% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCM_extract_region.R
\name{GCM_extract_region}
\alias{GCM_extract_region}
\title{Read subset and built a 5D array with netCDF GCM's files}
\usage{
GCM_extract_region(Path, Region, Period = "Historical", Var = "Pp")
}
\arguments{
\item{Path}{Directory where netCDF files are located}

\item{Region}{Region limits to subset data (MinLon, MaxLon, MinLat, MaxLat)}

\item{Period}{Select period to analysis- Could be 'Historical' or 'Future'. 'Historical' for default}

\item{Var}{Select the variable to process ('Pp', 'Tmax or 'Tmin'). 'Pp' data for default}
}
\value{
5D array with GCM data
}
\description{
Read subset and built a 5D array with netCDF GCM's files
}
