% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCM_spatial_down.R
\name{GCM_spatial_down}
\alias{GCM_spatial_down}
\title{Extract a downscaled GCM data for a point of interest}
\usage{
GCM_spatial_down(Input, RegLon, RegLat, XYcoords, YearIni)
}
\arguments{
\item{Input}{A 5D array (lon,lat,day,year,model) from GCM_extract_region output}

\item{RegLon}{Vector of longitudes for the study area}

\item{RegLat}{Vector of latitudes for the study area}

\item{XYcoords}{Vector with station coordinates (lon,lat)}

\item{YearIni}{Initial year to compute leap years}
}
\value{
GCM time series for a point of interest
}
\description{
Extract a downscaled GCM data for a point of interest
}
